pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-carrom"
version = "0.1.0"
description = "OpenEnv RL environment for Carrom"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "OpenEnv Carrom Contributors" }
]
keywords = ["reinforcement-learning", "openenv", "carrom", "pymunk"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

dependencies = [
  "numpy>=1.24",
  "pydantic>=2.0",
  "pymunk>=6.5",
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "openenv-core",
  "requests>=2.25",
  "matplotlib>=3.7",
  "Pillow>=9.0",
]

[project.optional-dependencies]
render = ["pygame>=2.5"]
train = [
  "trl>=0.12",
  "transformers>=4.40",
  "datasets>=2.18",
  "accelerate>=0.30",
  "peft>=0.10",
  "torch>=2.2",
]
dev = ["pytest>=7.4"]

[tool.setuptools.packages.find]
where = ["."]
include = ["carrom_env*"]
